refactor: 로또 번호 BonusBall로 포장
